Extract protocol health schemas
